[ADD] IWindow actualisation

+++ core/window/WindowModule.cpp
@@ -37,19 +37,19 @@ void WindowModule::clear() {
 	this->_object->clear();
 }
 
-void WindowModule::drawSprite(std::string asset, int color, std::pair<int, int> position) {
+void WindowModule::drawSprite(std::string asset, int color, std::pair<size_t, size_t> position) {
 	if (!this->_object)
 		throw WindowException("Error no lib Loaded");
 	this->_object->drawSprite(asset, color, position);
 }
 
-void WindowModule::drawRectangle(int color, std::pair<int, int> position) {
+void WindowModule::drawRectangle(int color, std::pair<size_t, size_t> position) {
 	if (!this->_object)
 		throw WindowException("Error no lib Loaded");
 	this->_object->drawRectangle(color, position);
 }
 
-void WindowModule::drawText(std::string text, int color, std::pair<int, int> position) {
+void WindowModule::drawText(std::string text, int color, std::pair<size_t, size_t> position) {
 	if (!this->_object)
 		throw WindowException("Error no lib Loaded");
 	this->_object->drawText(text, color, position);
